@@ -1,6 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, LOCALE_ID } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { RouterModule, PreloadAllModules } from '@angular/router';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms'
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations'
@@ -11,38 +11,44 @@ import {ROUTES} from './app.routes'
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
 import { HomeComponent } from './home/home.component';
-import { RestaurantsComponent } from './restaurants/restaurants.component';
-import { RestaurantComponent } from './restaurants/restaurant/restaurant.component';
-import { RestaurantDetailComponent } from './restaurant-detail/restaurant-detail.component';
-import { MenuComponent } from './restaurant-detail/menu/menu.component';
-import { ShoppingCartComponent } from './restaurant-detail/shopping-cart/shopping-cart.component';
-import { MenuItemComponent } from './restaurant-detail/menu-item/menu-item.component';
-import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
-
+import { CampanhasComponent } from './campanhas/campanhas.component';
+import { CampanhaComponent } from './campanhas/campanha/campanha.component';
+import { CampanhaDetalhesComponent } from './campanha-detalhes/campanha-detalhes.component';
+import { ProdutoComponent } from './campanha-detalhes/produto/produto.component';
+import { ShoppingCartComponent } from './campanha-detalhes/shopping-cart/shopping-cart.component';
+import { ProdutoCampanhaComponent } from './campanha-detalhes/produto-campanha/produto-campanha.component';
+import { ReviewsComponent } from './campanha-detalhes/reviews/reviews.component';
 import { OrderSummaryComponent } from './order-summary/order-summary.component';
 
-import {SharedModule} from './shared/shared.module';
-import { NotFoundComponent } from './not-found/not-found.component'
+import { SharedModule } from './shared/shared.module';
+import { NotFoundComponent } from './not-found/not-found.component';
+import { LoginComponent } from './security/login/login.component'
+import { UserDetailComponent } from './header/user-detail/user-detail.component';
+import { RegisterComponent } from './security/register/register.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     HeaderComponent,
     HomeComponent,
-    RestaurantsComponent,
-    RestaurantComponent,
-    RestaurantDetailComponent,
-    MenuComponent,
+    CampanhasComponent,
+    CampanhaComponent,
+    ProdutoCampanhaComponent,
+    ProdutoComponent,
     ShoppingCartComponent,
-    MenuItemComponent,
+    ProdutoCampanhaComponent,
     ReviewsComponent,
     OrderSummaryComponent,
-    NotFoundComponent
+    NotFoundComponent,
+    CampanhaDetalhesComponent,
+    LoginComponent,
+    UserDetailComponent,
+    RegisterComponent
    ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
-    HttpModule,
+    HttpClientModule,
     SharedModule.forRoot(),
     RouterModule.forRoot(ROUTES, {preloadingStrategy: PreloadAllModules})
   ],
