@@ -1,12 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import {FormGroup, FormBuilder, Validators, AbstractControl} from '@angular/forms'
-
 import {Router} from '@angular/router'
 
 import {RadioOption} from '../shared/radio/radio-option.model'
 import {OrderService} from './order.service'
-import {CartItem} from '../restaurant-detail/shopping-cart/cart-item.model'
-import {Order, OrderItem} from "./order.model"
+import {CartItem} from '../campanha-detalhes/shopping-cart/cart-item.model'
+
+
 
 @Component({
   selector: 'mt-order',
@@ -34,57 +34,39 @@ export class OrderComponent implements OnInit {
 
   ngOnInit() {
     this.orderForm = this.formBuilder.group({
-      name: this.formBuilder.control('', [Validators.required, Validators.minLength(5)]),
-      email: this.formBuilder.control('', [Validators.required, Validators.pattern(this.emailPattern)]),
-      emailConfirmation: this.formBuilder.control('', [Validators.required, Validators.pattern(this.emailPattern)]),
-      address: this.formBuilder.control('', [Validators.required, Validators.minLength(5)]),
-      number: this.formBuilder.control('', [Validators.required, Validators.pattern(this.numberPattern)]),
-      optionalAddress: this.formBuilder.control(''),
       paymentOption: this.formBuilder.control('', [Validators.required])
-    }, {validator: OrderComponent.equalsTo})
+    })
   }
 
-  static equalsTo(group: AbstractControl): {[key:string]: boolean} {
-    const email = group.get('email')
-    const emailConfirmation = group.get('emailConfirmation')
-    if(!email || !emailConfirmation){
-      return undefined
-    }
-    if(email.value !== emailConfirmation.value){
-      return {emailsNotMatch:true}
-    }
-    return undefined
-  }
-
-  itemsValue(): number {
-    return this.orderService.itemsValue()
+  cartItemsValue(): number {
+    return this.orderService.cartItemsValue()
   }
 
   cartItems(): CartItem[] {
     return this.orderService.cartItems()
   }
 
-  increaseQty(item: CartItem){
-    this.orderService.increaseQty(item)
+  increaseQty(produto: CartItem){
+    this.orderService.increaseQty(produto)
   }
 
-  decreaseQty(item: CartItem){
-    this.orderService.decreaseQty(item)
+  decreaseQty(produto: CartItem){
+    this.orderService.decreaseQty(produto)
   }
 
-  remove(item: CartItem){
-    this.orderService.remove(item)
+  remove(produto: CartItem){
+    this.orderService.remove(produto)
   }
 
-  checkOrder(order: Order){
-    order.orderItems = this.cartItems()
-      .map((item:CartItem)=>new OrderItem(item.quantity, item.menuItem.id))
-    this.orderService.checkOrder(order)
+  checkOrder(cartItems: CartItem[]){
+//    order.orderItems = this.cartItems()
+//      .map((produto:CartItem)=>new OrderItem(produto.quantity, produto.produtoCampanha.id))
+      this.orderService.checkOrder(cartItems)
       .subscribe( (orderId: string) => {
         this.router.navigate(['/order-summary'])
-        this.orderService.clear()
-    })
-    console.log(order)
+        this.orderService.clear()})
+    // })
+    // console.log(cartItems)
   }
 
 }
