@@ -1,15 +1,26 @@
+import { ProdutoCampanha } from 'app/campanha-detalhes/produto/produto-campanha.model';
+import {Pessoa} from '../pessoa/pessoa.model'
+
 class Order {
   constructor(
-    public address: string,
-    public number: number,
-    public optionalAddress: string,
-    public paymentOption: string,
-    public orderItems: OrderItem[] = []
+    public idCompra: number,
+    public idPessoa: Pessoa,
+    public idProdutoCampanha: ProdutoCampanha,
+    public qtd: number,
+    public estado: number,
+    public valor: number
   ){}
 }
 
-class OrderItem {
-  constructor(public quantity: number, public menuId: string){}
+class OrderPost {
+  constructor(
+    public idCompra: number,
+    public idPessoa: number,
+    public idProdutoCampanha: number,
+    public qtd: number,
+    public estado: number,
+    public valor: number
+  ){}
 }
 
-export {Order, OrderItem}
+export {Order, OrderPost}
