@@ -5,7 +5,8 @@ import {SharedModule} from '../shared/shared.module'
 
 import {OrderComponent} from './order.component'
 import {OrderItemsComponent} from './order-items/order-items.component'
-import {DeliveryCostsComponent} from './delivery-costs/delivery-costs.component'
+import {DeliveryCostsComponent} from './delivery-costs/delivery-costs.component';
+
 
 const ROUTES: Routes = [
   {path:'', component: OrderComponent}
