@@ -1,52 +1,65 @@
 import {Injectable} from '@angular/core'
 
-import {Http, Headers, RequestOptions} from '@angular/http'
+import {HttpClient, HttpHeaders, HttpParams} from '@angular/common/http'
 import {Observable} from 'rxjs/Observable'
 import 'rxjs/add/operator/map'
 
-import {ShoppingCartService} from '../restaurant-detail/shopping-cart/shopping-cart.service'
-import {CartItem} from '../restaurant-detail/shopping-cart/cart-item.model'
-import {Order, OrderItem} from './order.model'
+import {ShoppingCartService} from '../campanha-detalhes/shopping-cart/shopping-cart.service'
+import {CartItem} from '../campanha-detalhes/shopping-cart/cart-item.model'
+
 
 import {MEAT_API} from '../app.api'
 
+import {LoginService} from '../security/login/login.service'
+import { OrderPost } from './order.model'
+
+
 @Injectable()
 export class OrderService {
+  private orderPost: OrderPost[] = [];
 
-  constructor(private cartService: ShoppingCartService, private http: Http){}
+  constructor(private cartService: ShoppingCartService,
+    private http: HttpClient,
+    private loginService: LoginService
+  ){}
 
-  itemsValue(): number {
+  cartItemsValue(): number {
     return this.cartService.total()
   }
 
   cartItems(): CartItem[]{
-    return this.cartService.items
+    return this.cartService.carrinho
   }
 
-  increaseQty(item: CartItem){
-    this.cartService.increaseQty(item)
+  increaseQty(produto: CartItem){
+    this.cartService.increaseQty(produto)
   }
 
-  decreaseQty(item: CartItem){
-    this.cartService.decreaseQty(item)
+  decreaseQty(produto: CartItem){
+    this.cartService.decreaseQty(produto)
   }
 
-  remove(item: CartItem){
-    this.cartService.removeItem(item)
+  remove(produto: CartItem){
+    this.cartService.removeProduto(produto)
   }
 
   clear(){
     this.cartService.clear()
   }
 
-  checkOrder(order: Order): Observable<string> {
-    const headers = new Headers()
-    headers.append('Content-Type', 'application/json')
-    return this.http.post(`${MEAT_API}/orders`,
-                          JSON.stringify(order),
-                          new RequestOptions({headers: headers}))
-                    .map(response=> response.json())
-                    .map(order => order.id)
+  checkOrder(cartItems: CartItem[]): Observable<string> {
+
+    let headers = new HttpHeaders()
+    if (this.loginService.isLoggedIn()) {
+      headers = headers
+      .set('Authorization', `Bearer ${this.loginService.token.access}`)
+    }
+
+    return this.http.put<string>(`${MEAT_API}compraBulk/`,
+    cartItems, {headers:headers})
+
+    // return this.http.post<Order>(`${MEAT_API}orders`, order, {headers:headers})
+    //               .map(order => ' ') //order.address)
   }
 
 }
