@@ -5,5 +5,5 @@
 
 export const environment = {
   production: false,
-  api: 'http://localhost:3000'
+  api: 'http://127.0.0.1:8000/'
 };
